requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn>=0.29.0
pytest>=8.0.0
